requirements.txt
clickhouse==8.0.0
pandas==2.0.0
requests==2.31.0
beautifulsoup4==4.12.0
lxml==4.9.0
pyyaml==6.0
tqdm==4.65.0
